move elasticity to rheology model
