+++ include/aspect/material_model/viscoelastic.h
@@ -22,6 +22,7 @@
 #define _aspect_material_model_viscoelastic_h
 
 #include <aspect/material_model/interface.h>
+#include <aspect/material_model/rheology/elasticity.h>
 #include <aspect/simulator_access.h>
 #include <aspect/material_model/equation_of_state/multicomponent_incompressible.h>
 
@@ -30,27 +31,6 @@ namespace aspect
   namespace MaterialModel
   {
     using namespace dealii;
-
-    /**
-     * Additional output fields for the elastic shear modulus to be added to
-     * the MaterialModel::MaterialModelOutputs structure and filled in the
-     * MaterialModel::Interface::evaluate() function.
-     */
-    template <int dim>
-    class ElasticAdditionalOutputs : public NamedAdditionalMaterialOutputs<dim>
-    {
-      public:
-        ElasticAdditionalOutputs(const unsigned int n_points);
-
-        virtual std::vector<double> get_nth_output(const unsigned int idx) const;
-
-        /**
-         * Elastic shear moduli at the evaluation points passed to
-         * the instance of MaterialModel::Interface::evaluate() that fills
-         * the current object.
-         */
-        std::vector<double> elastic_shear_moduli;
-    };
 
     /**
      * An implementation of a simple linear viscoelastic rheology that only
@@ -245,10 +225,6 @@ namespace aspect
                                          const MaterialUtilities::CompositionalAveragingOperation &average_type) const;
 
 
-        double calculate_average_viscoelastic_viscosity (const double average_viscosity,
-                                                         const double average_elastic_shear_modulus,
-                                                         const double dte) const;
-
 
         EquationOfState::MulticomponentIncompressible<dim> equation_of_state;
 
@@ -263,32 +239,7 @@ namespace aspect
         std::vector<double> thermal_conductivities;
 
         /**
-         * Vector for field elastic shear moduli, read from parameter file.
-         */
-        std::vector<double> elastic_shear_moduli;
-
-        /**
-         * Bool indicating whether to use a fixed material time scale in the
-         * viscoelastic rheology for all time steps (if true) or to use the
-         * actual (variable) advection time step of the model (if false). Read
-         * from parameter file.
-         */
-        bool use_fixed_elastic_time_step;
-
-        /**
-         * Bool indicating whether to use a stress averaging scheme to account
-         * for differences between the numerical and fixed elastic time step
-         * (if true). When set to false, the viscoelastic stresses are not
-         * modified to account for differences between the viscoelastic time
-         * step and the numerical time step. Read from parameter file.
-         */
-        bool use_stress_averaging;
-
-        /**
-         * Double for fixed elastic time step value, read from parameter file
-         */
-        double fixed_elastic_time_step;
-
+        Rheology::Elasticity<dim> elastic_rheology;
     };
 
   }
