+++ include/aspect/material_model/viscoelastic_plastic.h
@@ -23,7 +23,7 @@
 
 #include <aspect/material_model/interface.h>
 #include <aspect/simulator_access.h>
-#include <aspect/material_model/viscoelastic.h>
+#include <aspect/material_model/rheology/elasticity.h>
 #include <aspect/material_model/equation_of_state/multicomponent_incompressible.h>
 
 namespace aspect
